Rangeify main decode process (simplifying)
The previous executeDecodeProcess() function was extremely hard to
follow, so we replace the main path with a pair of ranges which
decode external files: first parsing NMEA lines, then whole AIS
messages.

In future we're going to need to read more than just AIS NMEA files,
so we follow a similar interfaces-based strategy as we did for file
writers.

README.md will need updating in line with these changes.